power: qpnp-fg-gen3: queue work on system_power_efficient_wq
There doesn't seem to be any real dependency of scheduling these on
the cpu which scheduled them, so moving every *_delayed to the
power efficient wq save potential needlessly idle cpu wake ups
leaving the scheduler to decide the most appropriate cpus to wake up.

These are called very often so I wouldn't be susprised if the power
savings with this change were somewhat significant.

Change-Id: I23ef589adb1f18818bb25471bf98b82a36416496
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>