cpu_input_boost: Rework scheduling setup
This patch minimizes the latency of an input boost by creating a
dedicated kworker thread and scheduling it with the SCHED_FIFO realtime
policy. We use an RT priority of 98 (max = 100) to have the input boost
task preempt userspace RT tasks if needed while being preempted by
in-kernel RT tasks which are set at priority 99 or higher. To reduce
impact on the overall system performance due to the frequent preemption,
we bind the kworker to only run on last 3 little cluster CPU cores. CPU
0 is considered too busy, so leave the boosting workers off of that
core.

Also, since now the cpu_input_boost workqueue would only handle the work
to disable an active input boost, we can remove the workqueue entirely
and move the disabling work to the shared system-wide power efficient
workqueue because it's not latency critical.

Thanks to TheCrazyLex (Alex Naidis <alex.naidis@linux.com>) for the
original version of this patch for the cpu-boost driver; I simply
converted it over to cpu_input_boost.

Sources used:
 - https://github.com/arter97/android_kernel_oneplus_sdm845/commit/a2790f17a500b9abcc070cc6438646ce5245779a
 - Razer Phone 2 "aura" kernel source

*adapted to current model [spak]

Signed-off-by: kdrag0n <dragon@khronodragon.com>