cpu-boost: Remove unnecessary stune boost reset at the start of boost
The intention behind resetting stune boost at the start of every boost
was to ensure that changes to dynamic_stune_boost take effect as soon as
possible. However, the boost reset seems like a poor way to tackle the
issue since the dynamic_stune_boost value will likely not be changing
for the majority of the time. It would be better to reset the stune
boost only if the newest dynamic_stune_boost value differs from the
previous one.