Revert "sched: fix task-priority when using SCHED_RESET_ON_FORK"
This reverts commit 8c36cce3d4102e5f017c2bc1d70a3a952634c315.

SCHED_RESET_ON_FORK resetting the scheduling policy of a task from an RT
policy such as SCHED_FIFO to SCHED_NORMAL is intended behavior. This
commit breaks that behavior.

Mainline doesn't have this commit or an equivalent change as far as I
can tell, so this should be safe to revert.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>